+++ package.json
@@ -10,6 +10,6 @@
   },
   "scripts": {
     "build:css": "tailwindcss -i ./app/assets/stylesheets/application.tailwind.css -o ./app/assets/builds/application.css --minify",
-    "build": "esbuild app/javascript/*.* --bundle --sourcemap --outdir=app/assets/builds --public-path=assets"
+    "build": "esbuild app/javascript/controllers/*.* --bundle --sourcemap --outdir=app/assets/builds --public-path=assets"
   }
 }
